Refactor to smaller files with tests
